support XDG_STATE_HOME
There is a proposal to add another directory type to the XDG Base Directory
Specification.

https://wiki.debian.org/XDGBaseDirectorySpecification#state

To be clear, this is _not_ currently part of the specification.  However, I
think it would be a valuable addition to the appdirs module.  I wasn't sure if
there was an equivalent for OSX or Windows, so for now those fall back to using
user_data_dir.